requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pandas==2.1.4
numpy==1.26.2
mangum==0.17.0
pydantic==2.5.0
python-multipart==0.0.6
pyyaml==6.0.1
requests==2.31.0
python-dotenv==1.0.0
httpx==0.25.2 